deck-build-log/deck-build-log-plot: Also show success
To make it easier to distinguish between "everything else is good" or
"this is everything, and we're completely broken" when you don't see a
lot of non-matching failures in a given time range.  To support this,
I'm saving the JSON with all the jobs regardless of state (but still
filtered by job name) into $CACHE/jobs.json.  I've also transitioned
to using a legend to explain the colors, although (at least in my RHEL
7.5's ancient, stock v1.2) that requires a bit of hoop-jumping.

To reflect the changes, I've also rebuild the examples to show the
us-east-1b RunInstances outage AWS just surprised us with [1] ;).

[1]: https://github.com/openshift/release/pull/3204