vpcs/calls-per-vpc: Fix calls-per-VPC subplot
The previous approach was confused by the stacked-ness of the n
returned by hist().  This commit still uses the bins from hist(), but
calculates the counts directly to avoid confusion.

I'm using both >= and <= (instead of, say >= and <), because I expect
timestamps to only exactly fall on a bin edge for the first and last
bin.  If that assumption breaks down, it's only going to be an
off-by-one anyway.