agents-and-events: Inline SQL and track API errors
Inlining the SQL makes updating the Lambda function easier.  We'd have
to bump it anyway for a new query ID, so with an inline script there's
one less thing to bump (just the script) vs. two (the script and the
saved query).

Also split out errors.  This makes it easier to find broken code, and
avoids confusing me when I try to figure out why there are more
CreateNatGateway calls than I expect, because the numbers no longer
conflate successful calls (which I have a handle on) with failed calls
(which can vary wildly depending on overall account activity).