/deck-build-log/deck-build-log-plot: Print pull-request links
Sorted by error message counts, and then within each error message, by
pull-request counts, with links to the affected pull request.  This
makes it easy to see who to talk to, in a form that you can paste into
chat.  It also makes it easy to distinguish between issues that affect
the org as a whole, vs. issues localized to a particular team.