vpcs/calls-per-vpc: Log y-scale and error-code breakdown
Because sometimes it's useful to distinguish between different kinds
of errors, and sometimes you want to see small values that would be
swamped by something larger on a linear scale.  error_suffix and
yscale should be wired up to an argparse parser or some such, but I
haven't bothered yet.